requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==4.2
django-ninja==1.1.0
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==8.1.1
pytest-django==4.8.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
